requirements.txt
anthropic==0.45.2
firebase_admin==6.6.0
Flask==3.1.0
flask_cors==5.0.0
protobuf==5.29.3
pymongo==4.11
python_bcrypt==0.3.2
Requests==2.32.3
google-ai-generativelanguage==0.6.15
google-api-core==2.24.1
google-api-python-client==2.160.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.0.0
google-generativeai==0.8.4
google-pasta==0.2.0
googleapis-common-protos==1.66.0
grpcio==1.67.1
grpcio-status==1.67.1
google-crc32c==1.6.0
itsdangerous==2.2.0
python-dotenv==1.0.1
msgpack==1.1.0
google-cloud-core==2.4.1
cryptography==44.0.0
google-cloud-firestore==2.20.0
click==8.1.8
google-resumable-media==2.7.2
CacheControl==0.14.2
firebase-admin==6.6.0
google-cloud-storage==3.0.0
python-bcrypt==0.3.2
blinker==1.9.0
dnspython==2.7.0
Flask-Cors==5.0.0
firebase_admin==6.6.0
ez_setup==0.9